\name{compCol}
\alias{compCol}
\title{Percent of complete rows}
\usage{
  compCol(df)
}
\arguments{
  \item{df}{data frame under consideration}
}
\value{
  data.frame
}
\description{
  This displays the percentage of
  \code{\link{complete.cases}}
}
\examples{
data(orangeCar.test)
compRow(orangeCar.test)
}
\author{
  Kenny Darrell
}
\keyword{info}

