% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{CV}
\alias{CV}
\title{Cross-Validation with Penalized Likelihood Chronograms}
\usage{
CV(
  phy,
  lambda = 10^(-3:2),
  model = "correlated",
  quiet = FALSE,
  calibration = makeChronosCalib(phy)
)
}
\arguments{
\item{phy}{an object of class \code{"phylo"}.}

\item{lambda}{a numeric vector with the values of smoothing
parameter to be assessed.}

\item{model}{the model of rate evolution. currently set to autocorrelated.}

\item{quiet}{a logical value.}

\item{calibration}{a data frame with the calibration dates of the
chronogram.}
}
\value{
a matrix with two columns named \code{lambda} and
\code{CV}.
}
\description{
Computes the cross-validation (CV) of a chronogram
using penalized likelihood.
}
\details{
This follows the CV calculation described in Sanderson (2002).
}
\examples{
## a simple random tree, so lambda is expected to be zero
library(ape)
tr <- rtree(10)
res <- CV(tr, 10^seq(-4, 2, 0.25))
plot(res, type = "o", log = "xy")

## create 'artificial' auto-correlation among branches by sorting
## them according to their lengths:
tr$edge.length <- sort(tr$edge.length)
res <- CV(tr, 10^seq(-4, 2, 0.25))
plot(res, type = "o", log = "xy")
}
\references{
Paradis, E., Claramunt, S., Brown, J., and Schliep,
K. Confidence intervals in molecular dating by penalized
likelihood. (in preparation)

Sanderson, M. J. (2002) Estimating absolute rates of
molecular evolution and divergence times: a penalized
likelihood approach. \emph{Molecular Biology and Evolution},
\bold{19}, 101--109.
}
\seealso{
\code{\link[=qage]{qage()}} \code{\link[=drawChronosCI]{drawChronosCI()}} \code{\link[ape:chronos]{ape::chronos()}} \code{\link[ape:chronos]{ape::makeChronosCalib()}}
}
\author{
Emmanuel Paradis, Santiago Claramunt, Joseph Brown, Klaus Schliep
}
\keyword{models}
