% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qage.R
\name{qage}
\alias{qage}
\title{Quantile-based age estimation}
\usage{
qage(p = 0.5, ages, method = "StraussSadler", k = 5)
}
\arguments{
\item{p}{The desired probability level (0 < p < 1). A vector of
probabilities can be provided. Default p = 0.5}

\item{ages}{Either a vector of fossil ages or a matrix with two columns: the
first with the minimum age bounds (upper stratigraphic bounds) and the
second with the maximum age bounds (lower stratigraphic bounds) of each
fossil (in rows). A minimum of 2 are required}

\item{method}{The method for modelling age uncertainty. A number of options
is available:
\itemize{
\item{\code{"StraussSadler"}} {The (default) method of Strauss & Sadler
(1989) assumes that the sampled fossil ages are uniformly distributed in
time, and a warning is returned if a Kolmogorov-Smirnov test rejects the
uniformity hypothesis.}
\item{\code{“Beta”}} {The method is a different parameterization of the
Strauss & Sadler method (Wang et al., 2009) that uses the qbeta function,
since the ratio between the observed maximum fossil age and the clade age
for the Strauss & Sadler model is distributed according to a Beta
distribution with parameters N and 1 (Wang et al. 2009); this should give
the same result as "StraussSadler".}
\item{\code{Solow}} {The method of Solow (2003) does not assume uniformity
in sampled fossil ages, but is based on the two oldest ages only.}
\item{\code{NorrisPenG} or \code{NorrisGLin}} {The method of Norris et al.
(2015) based on the two oldest fossils only and the log-logistic
distribution. "NorrisPenG" is used when the precise phylogenetic placement
of fossils is not known, whereas "NorrisGLin" is used when one fossil from
each daughter lineage is used.}
\item{\code{"RobertsSolow"}} {The method of Roberts & Solow (2003) does not
assume uniformity in sampling and is based on the fact that the joint
distribution of the k oldest fossil ages can be modeled with the same
Weibull distribution.}
}}

\item{k}{The number of fossil ages to use in when
\code{method="RobertsSolow"}, in which only the k oldest fossils are used.
Default k = 5.}
}
\value{
A numeric value (or vector of numeric values, if multiple p values
are provided) representing the age estimate of the clade origin given the
method a p value provided
}
\description{
Quantile function that computes the age corresponding to a
particular probability for the upper bound of a distribution of ages
(Strauss & Sadler 1989, Gingerich & Uhen 1998, Solow 2003). The method of
Strauss & Sadler (1989) assumes that the distribution of fossil ages is
uniform and their formula depends on the fossil ages range and the number of
fossil ages. The method of Solow (2003) is a general method for non-uniform
distributions and depends on the temporal gap between the oldest and the
second oldest fossil ages. Both methods assume that fossil ages are
independent samples from the same distribution (only relevant for the two
oldest ages for Solow's method), therefore, fossils should be as independent
as possible (ideally from different geological formations and different
regions).

In the particular case where there are only two fossil ages, Strauss &
Sadler's and Solow's methods converge to the same result; the quantile
functions are simply Xn/(1-P), and the likelihood function is 1/X.
}
\examples{
\dontrun{
  # The following demonstrates how inferences depend on p and method
  qage(p=c(0.1, 0.5, 0.9), ages=c(54, 30, 25, 14, 5))
  qage(p=c(0.1, 0.5, 0.9), ages=c(54, 30, 25, 14, 5), method="Beta")
  qage(p=c(0.1, 0.5, 0.9), ages=c(54, 30, 25, 14, 5), method="Solow")
  }
}
\references{
\insertRef{Gingerich1998}{chronos}

\insertRef{Norris2015}{chronos}

\insertRef{Solow2003}{chronos}

\insertRef{Strauss1989}{chronos}

\insertRef{Wang2007}{chronos}

\insertRef{Wang2009}{chronos}

\insertRef{Wang2010}{chronos}
}
\keyword{models}
