\name{readTODO}
\alias{readTODO}
\title{Display contents of todo.txt}
\usage{
  readTODO()
}
\arguments{
  \item{NULL}{}
}
\value{
  data.frame
}
\description{
  This function will show you what is contained in the
  todo.txt file that is created from use of the
  \code{\link{TODO}} function.
}
\examples{
TODO()
readTODO()
}
\author{
  Kenny Darrell
}
\keyword{tools}

