\name{sax}
\alias{sax}
\title{sax}
\usage{
  sax(ts, bp = 3, n = 10, plot = FALSE)
}
\arguments{
  \item{ts}{The time series to condense.}

  \item{bp}{The number of piecewise sections to generate.}

  \item{n}{fsd}

  \item{plot}{fsdf}
}
\description{
  This function condenses
}
\examples{
x.Date <- as.Date("2003-02-01") + c(1, 3, 7, 9, 14, 18) - 1
x <- xts(rnorm(6), x.Date)
paa(x, 2)
}
\keyword{sax}
\keyword{timeseries,}

