\name{timer}
\alias{timer}
\title{Time functions}
\usage{
  timer(fun, iters, ...)
}
\arguments{
  \item{df}{data frame under consideration}
}
\value{
  data.frame
}
\description{
  Get the average time for a function call.
}
\examples{
data(orangeCar.test)
compRow(orangeCar.test)
}
\author{
  Kenny Darrell
}
\keyword{info}

