\name{TODO}
\alias{TODO}
\title{Capture items which needs more work}
\usage{
  TODO()
}
\arguments{
  \item{NULL}{Ran only for side effects.}
}
\value{
  NULL Ran only for side effects.
}
\description{
  This function will parse all of the code files and add
  any mention of the todo tag to a file called todo.txt.
  This can be used for adding notes to . for the future.
}
\examples{
TODO()
readTODO()
}
\author{
  Kenny Darrell
}
\keyword{tools}

