% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronosCI.R
\name{chronosCI}
\alias{chronosCI}
\title{Confidence Intervals of Chronograms}
\usage{
chronosCI(
  chronogram,
  pml.output,
  B = 100,
  type = "semiparametric",
  calibration = NULL,
  method = "StraussSadler",
  control = NULL,
  quiet = FALSE,
  trees = FALSE
)
}
\arguments{
\item{chronogram}{a chronogram output by \code{\link[ape:chronos]{ape::chronos()}}.}

\item{pml.output}{an unrooted tree output by \code{\link[phangorn:pml]{phangorn::optim.pml()}}.}

\item{B}{the number of replications.}

\item{type}{the bootstrap method to be used. Can be either a character
string among "nonparametric", "semiparametric", or "parametric", or an
integer between 1 and 3.}

\item{calibration}{a data frame with the calibration points of the tree.}

\item{method}{a character string specifying the method used to assess
calibration uncertainty. See \code{\link[=qage]{qage()}} for available methods.}

\item{control}{see \code{\link[ape:chronos]{ape::chronos()}}. By default, these are the same than
used to estimate the chronogram with \code{\link[=chronos]{chronos()}}.}

\item{quiet}{a logical value. By default, the progress of the
computations is printed.}

\item{trees}{a logical value specifying whether to return the bootstrap
trees produced by phangorn (FALSE, by default).}
}
\value{
a matrix with the 50\\% and 95\\% lower and upper bounds of the
confidence intervals.
}
\description{
Computes the confidence intervals of a chronogram using
different methods.
}
\details{
The details of the methods are presented in the manuscript below.

The labels (or taxa names) of the first argument (\code{chronogram})
must be present in the second argument (\code{pml.output}), and this
second one must also include the outgroup.
}
\examples{
\dontrun{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)for the standard data sets.
}
}
\references{
Paradis, E., Claramunt, S., Brown, J., and Schliep, K. Confidence
intervals in molecular dating by penalized likelihood. (in preparation)
}
\seealso{
\code{\link[=qage]{qage()}} \code{\link[=drawChronosCI]{drawChronosCI()}} \code{\link[ape:chronos]{ape::chronos()}}
\code{\link[phangorn:bootstrap.pml]{phangorn::bootstrap.pml()}}
}
\author{
Emmanuel Paradis, Santiago Claramunt, Joseph Brown, Klaus Schliep
}
\keyword{models}
