% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronosCI.R
\name{chronosCI}
\alias{chronosCI}
\title{Confidence Intervals of Chronograms}
\usage{
chronosCI(
  chronogram,
  pml.output,
  B = 100,
  smoothRates = TRUE,
  Poisson = FALSE,
  calibration = NULL,
  method = "StraussSadler",
  quiet = FALSE
)
}
\arguments{
\item{chronogram}{a chronogram output by \code{\link[ape]{chronos}}.}

\item{pml.output}{an unrooted tree output by
\code{\link[phangorn]{optim.pml}}.}

\item{B}{the number of replications.}

\item{smoothRates}{a logical value: should the distributions of
substitution rates be smoothed? (Ignored if the \code{Poisson = TRUE}.)}

\item{Poisson}{a logical value: should samples of branch lengths be
generated by Poisson simulation?}

\item{calibration}{a data frame with the calibration points of the tree.}

\item{method}{a character string specifying the method used to assess
calibration uncertainty.}

\item{quiet}{a logical value. By default, the progress of the
computations is printed.}
}
\value{
a matrix with the 50\% and 95\% lower and upper bounds of the
confidence intervals.
}
\description{
Computes the confidence intervals of a chronogram using
different methods.
}
\details{
The details of the methods are presented in the manuscript below.
\cr\cr
The labels (or taxa names) of the first argument (\code{chronogram})
must be present in the second argument (\code{pml.output}), and this
second one must also include the outgroup.
}
\examples{
\dontrun{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)for the standard data sets.
}
}
\references{
Paradis, E., Claramunt, S., Brown, J., and Schliep, K. Confidence
intervals in molecular dating by penalized likelihood. (in preparation)
}
\seealso{
\code{\link{drawChronosCI}}
}
\author{
Emmanuel Paradis, Santiago Claramunt, Joseph Brown, Klaus Schliep
}
\keyword{models}
