% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawChronosCI.R
\name{drawChronosCI}
\alias{drawChronosCI}
\title{Draw Confidences Intervals on Phylogenies}
\usage{
drawChronosCI(
  CI,
  col95 = "#FF00004D",
  col50 = "#0000FF4D",
  height = NULL,
  legend = TRUE,
  ...
)
}
\arguments{
\item{CI}{output from \code{\link{chronosCI}} or a similar matrix.}

\item{col95}{colour used for the 95\% intervals; by default: transparent
red.}

\item{col50}{colour used for the 50\% intervals; by default: transparent
blue.}

\item{height}{the height of the boxes.}

\item{legend}{a logical value.}

\item{\dots}{arguments passed to \code{\link[graphics]{legend}}}
}
\description{
This is a low-level plotting command to draw the confidence
intervals on the node of a tree as rectangles with coloured backgrounds.
}
\details{
The matrix \code{CI} must have four rows and as many columns as the
number of nodes of the tree. The first and fourth rows give the lower and
upper bounds of the 95\% confidence intervals. The second and third rows
give the lower and upper bounds of the 50\% confidence intervals.
}
\examples{
\dontrun{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)for the standard data sets.
}
}
\seealso{
\code{\link{chronosCI}}
}
\author{
Emmanuel Paradis, Santiago Claramunt, Joseph Brown, Klaus Schliep
}
\keyword{aplot}
